+++ crates/core/src/config/mod.rs
@@ -73,6 +73,9 @@ pub struct Config {
     pub compression_format: CompressionFormat,
 
     #[serde(default)]
+    pub zstd_compression_level: i32,
+
+    #[serde(default)]
     #[cfg(target_os = "windows")]
     pub close_to_tray: bool,
 
@@ -369,6 +372,7 @@ pub async fn load_config() -> Result<Config, FilesystemError> {
 const fn default_true() -> bool {
     true
 }
+
 #[cfg(test)]
 mod test {
     use super::*;